Dynamic parachute assignment for Fallschirmjäger
Fallschirmjägers now have a variable at the top of their file to control what type of backpack they spawn with: a parachute or regular backpack. FSJP files are now no longer needed.

Also includes a small change Feld made that was supposed to be in the US derandomisation PR.