Remove Zeus from supporting fires module (#117)
Removes the Zeus option from the supporting fires module as it was never used and causes issues when the platoon commander is also logged in as admin (giving him Zeus). Being Zeus would give the platoon commander infinite ammo and perfect accuracy.

(I want to have it mentioned for the record, I DID fix the initial bug in the Zeus only code, but then decided to just axe the entire Zeus option from the module because of the above-mentioned issue. But I DID fix it!)